+++ DaddysPlaceApi/Entity/PaymentEntity.cs
@@ -3,8 +3,8 @@
     public class PaymentEntity
     {
         public int Id { get; set; }
-        public float Receive { get; set; }
-        public float Refund { get; set; }
+        public double Receive { get; set; }
+        public double Refund { get; set; }
         public int Type { get; set; }
         public int Frn_OrderId { get; set; }
     }
