+++ DaddysPlaceApi/ViewEntity/ItemViewEntity.cs
@@ -3,9 +3,9 @@
     public class ItemViewEntity
     {
         public int Id { get; set; }
-        public float Discount { get; set; }
-        public float Quanlity { get; set; }
-        public float? Price { get; set; }
+        public decimal Discount { get; set; }
+        public decimal Quanlity { get; set; }
+        public decimal? Price { get; set; }
         public int Frn_ProductId { get; set; }
         public int Frn_OrderId { get; set; }
     }
